Update import for env.js file.
Updated script tag to use absolute path for the import of env.js. Previously, any path that was not in the root folder would not be able to load env.js. For example, if a user routed to host.com/home, env.js would be found because it lives in the root folder. But, if a user routed to host.com/users/1, env.js could not be found because it does not live at host.com/users/env.js.